Updated sequence generation.
Now each segment in a step has its own 'reservoir' of patterns.
E.g. if each step has segments ABBB, then we would start with one reservoir
containing 30 patterns of category A, and three reservoirs, each containing
patterns of category B.

As we are going over the sequence and randomly picking patterns, we always pick
from the reservoir of the corresponding segment and we don't "put the
pattern back".

If we run out of patterns in a particular reservoir, we reset it to
contain all 30 patterns as at the begnining.

This makes sure each of the patterns is played equaly often in each
segment.

I also updated the parameters (as we discussed on Friday), so the experiment
is counterbalanced.