On-line response logging
I have moved logging into the fast loop that collects tapping. I did
this for both, the "old" logging and BIDS.
This way, the information about the tap is saved into text file as soon
as the tap is made. Therefore if the experiment is terminated during a
sequence we don't lose data.

I have also modified the path to save the BIDS log, because it was
saving outside the main folder.

I also renamed the field currSeq.onset to currSeq.patternOnset, I find
it more informative.