Amplitude logging
I fixed a small thing in makeSequence.m that was from my previous commit
and caused a crash after Ceren's changes with ISO loudness equalization.

I also propose to log the acual matlab tone amplitudes that we're using
for each pattern, as opposed to logging the gain factor that we use to
multiply the baseAmp. The gains can be easily recalculated knowing the
baseAmp value, and this way we can directly see what the matlab
amplitudes are (e.e. that there is no clipping).