Adding non-repeating Grahn2007 sequences
Also big update of instructions. Now there is possibility to display
multiple instruction pages at the begining of the tapTrainer session.
Also, each sequence (pattern/audiotrack/non-repeating sequence) has 4
instruction files associated with it:
before          presented before the trial starts
during1         presented in the first part of the trial (when cue is
                silent)
during2         presnted in the second part of the trial (when the cue
                is added)
after           presented after the sound stops

Also update of stimulus-download. Now the list of required files is
saved as text file REQUIRED_FILES_LIST in the ./stimuli folder. The
script checks if all files listed there are available in the stimuli
folder. If not, it will download from Dropbox.