MainExp instructions update
Big update of instructions in the main experiment. Now we can create
multiple text files with name instrMainExpIntro# to be sequentially
displayed at the begining of the experiment. Also, after each sequence
there is a possibility to toggle general instructions for the
experiment, saved in the file instrMainExpGeneral. Also, there is a
possibility to display additional instructions after sequence #,
specified in a text file named instrmainExpDelay# (# is the seqi after
which the instruction is presented).

Also, README update.