Updated tone ramp parameters
There was an error in parameters controlling onset/offest ramps for the
tones in main experiment. Now the values are fixed to 10ms onset ramp,
20ms offset ramp.

Also, originally we were defining these as proportion of gridIOI, but
after thinking about it, I decided to change these parameters into
fixed values in seconds.

This is because if we decided to use very long gridIOIs in the future,
it could lead to very long-duration tones without clear onsets, which is
undesirable.